+++ tests/functional/test_gv.py
@@ -20,7 +20,7 @@ from NeoVintageous.tests import unittest
 
 class Test_gv(unittest.FunctionalTestCase):
 
-    def test_gv_works_for_visual_mode(self):
+    def test_v_gv(self):
         self.visual('a|b\nc|d\n')
         self.feed('<Esc>')
         self.assertNormal('ab\n|cd\n')
@@ -28,23 +28,51 @@ class Test_gv(unittest.FunctionalTestCase):
         self.assertVisual('a|b\nc|d\n')
         self.assertStatusLineRegex('-- VISUAL --')
 
-    def test_gv_works_for_visual_line_mode(self):
+    def test_l_gv(self):
         self.vline('ab\n|cd|\ne\n')
         self.feed('<Esc>')
         self.assertNormal('ab\nc|d\ne\n')
         self.feed('gv')
-        self.assertVline('ab\n|cd|\ne\n')
+        self.assertVline('ab\n|cd\n|e\n')
         self.assertStatusLineRegex('-- VISUAL LINE --')
 
-    def test_gv_works_for_visual_lines_mode(self):
+    def test_l_gv_many_lines(self):
         self.vline('ab\n|cd\nef|\ng\n')
         self.feed('<Esc>')
         self.assertNormal('ab\ncd\ne|f\ng\n')
         self.feed('gv')
-        self.assertVline('ab\n|cd\nef|\ng\n')
+        self.assertVline('ab\n|cd\nef\n|g\n')
         self.assertStatusLineRegex('-- VISUAL LINE --')
 
-    def test_gv_works_for_visual_block_mode(self):
+    def test_l_gv_empty_line(self):
+        self.vline('x\n|\n\n\n|\n\ny')
+        self.feed('<Esc>')
+        self.feed('gv')
+        self.assertVline('x\n|\n\n\n|\n\ny')
+        self.assertStatusLineRegex('-- VISUAL LINE --')
+
+    def test_l_gv_eof(self):
+        self.vline('1\n|2\n3\ny|')
+        self.feed('<Esc>')
+        self.feed('gv')
+        self.assertVline('1\n|2\n3\ny|')
+        self.assertStatusLineRegex('-- VISUAL LINE --')
+
+    def test_l_gv_reverse(self):
+        self.rvline('1\n|2\n3\n|xy')
+        self.feed('<Esc>')
+        self.feed('gv')
+        self.assertRVline('1\n|2\n3\n|xy')
+        self.assertStatusLineRegex('-- VISUAL LINE --')
+
+    def test_l_gv_reverse_eof(self):
+        self.rvline('1\n|2\n3\ny|')
+        self.feed('<Esc>')
+        self.feed('gv')
+        self.assertRVline('1\n|2\n3\ny|')
+        self.assertStatusLineRegex('-- VISUAL LINE --')
+
+    def test_b_gv(self):
         self.vblock('a\nb|cd|e\nf|gh|i\nj\n')
         self.feed('<Esc>')
         self.assertNormal('a\nbcde\nfg|hi\nj\n')
@@ -52,10 +80,39 @@ class Test_gv(unittest.FunctionalTestCase):
         self.assertVblock('a\nb|cd|e\nf|gh|i\nj\n')
         self.assertStatusLineRegex('-- VISUAL BLOCK --')
 
+    def test_noop(self):
+        self.normal('a|bc')
+        self.view.erase_regions('visual_sel')
+        self.view.settings().erase('_nv_visual_sel_mode')
+        self.feed('gv')
+        self.assertNormal('a|bc')
+        self.assertStatusLineRegex('')
+
     def test_issue_338(self):
         self.vline('1111\n2222\n    |3333\n    4444\n|')
         self.feed('<Esc>')
         self.assertNormal('1111\n2222\n    3333\n    444|4\n')
         self.feed('gv')
         self.assertVline('1111\n2222\n|    3333\n    4444\n|')
+        self.assertStatusLineRegex('-- VISUAL LINE --')
+
+    def test_issue_426_l_eol_newline(self):
+        self.vline('111\n|222\n333\n|\n5')
+        self.feed('<Esc>')
+        self.feed('gv')
+        self.assertVline('111\n|222\n333\n|\n5')
+        self.assertStatusLineRegex('-- VISUAL LINE --')
+
+    def test_issue_426_v(self):
+        self.visual('111\na|bc\nx|yz\n444\n5')
+        self.feed('<Esc>')
+        self.feed('gv')
+        self.assertVisual('111\na|bc\nx|yz\n444\n5')
+        self.assertStatusLineRegex('-- VISUAL --')
+
+    def test_issue_426_l(self):
+        self.vline('111\n|222\n333\n|444\n5')
+        self.feed('<Esc>')
+        self.feed('gv')
+        self.assertVline('111\n|222\n333\n|444\n5')
         self.assertStatusLineRegex('-- VISUAL LINE --')